Combine debugStackTraceTracker and debugLocationTracker
