+++ src/client/datascience/debugLocationTracker.ts
@@ -3,19 +3,19 @@
 'use strict';
 import { DebugAdapterTracker, Event, EventEmitter } from 'vscode';
 import { DebugProtocol } from 'vscode-debugprotocol';
-import { DebugStackTraceTracker } from './jupyter/debugStackTraceTracker';
 
 import { IDebugLocation } from './types';
 
 // When a python debugging session is active keep track of the current debug location
-export class DebugLocationTracker extends DebugStackTraceTracker implements DebugAdapterTracker {
-    private waitingForStackTrace: boolean = false;
+export class DebugLocationTracker implements DebugAdapterTracker {
+    protected topMostFrameId = 0;
+    protected stackFrameRequestSequenceNumber: number = -1; // Keep track of the sequence number
+    private waitingForStackTrace = false;
     private _debugLocation: IDebugLocation | undefined;
     private debugLocationUpdatedEvent: EventEmitter<void> = new EventEmitter<void>();
     private sessionEndedEmitter: EventEmitter<DebugLocationTracker> = new EventEmitter<DebugLocationTracker>();
 
-    constructor(private _sessionId: string) {
-        super();
+    constructor(private _sessionId: string | undefined) {
         this.DebugLocation = undefined;
     }
 
@@ -35,9 +35,21 @@ export class DebugLocationTracker extends DebugStackTraceTracker implements Debu
         return this._debugLocation;
     }
 
-    // tslint:disable-next-line:no-any
     public onDidSendMessage(message: DebugProtocol.Response) {
-        super.onDidSendMessage(message);
+        if (this.isResponseForRequestToFetchAllFrames(message)) {
+            // This should be the top frame. We need to use this to compute the value of a variable
+            const topMostFrame = message.body.stackFrames[0];
+            this.topMostFrameId = topMostFrame?.id;
+            // If we are waiting for a stack trace, check our messages for one
+            if (this.waitingForStackTrace) {
+                this.DebugLocation = {
+                    lineNumber: topMostFrame?.line,
+                    fileName: this.normalizeFilePath(topMostFrame?.source?.path),
+                    column: topMostFrame.column
+                };
+                this.waitingForStackTrace = false;
+            }
+        }
         if (this.isStopEvent(message)) {
             // Some type of stop, wait to see our next stack trace to find our location
             this.waitingForStackTrace = true;
@@ -48,19 +60,17 @@ export class DebugLocationTracker extends DebugStackTraceTracker implements Debu
             this.DebugLocation = undefined;
             this.waitingForStackTrace = false;
         }
-
-        if (this.waitingForStackTrace) {
-            // If we are waiting for a stack track, check our messages for one
-            const debugLoc = this.getStackTrace(message);
-            if (debugLoc) {
-                this.DebugLocation = debugLoc;
-                this.waitingForStackTrace = false;
-            }
-        }
     }
 
-    public onWillStopSession() {
-        this.sessionEndedEmitter.fire(this);
+    public onWillReceiveMessage(message: DebugProtocol.Request) {
+        if (message.type === 'request' && message.command === 'stackTrace' && message.arguments.startFrame === 0) {
+            // VSCode sometimes sends multiple stackTrace requests. The true topmost frame is determined
+            // based on the response to a stackTrace request where the startFrame is 0 (i.e. this request
+            // retrieves all frames). Here, remember the sequence number of the outgoing request whose
+            // startFrame === 0, and update this.topMostFrameId only when we receive the response with a
+            // matching sequence number.
+            this.stackFrameRequestSequenceNumber = message.seq;
+        }
     }
 
     // Set our new location and fire our debug event
@@ -73,7 +83,15 @@ export class DebugLocationTracker extends DebugStackTraceTracker implements Debu
         }
     }
 
-    // tslint:disable-next-line:no-any
+    private normalizeFilePath(path: string): string {
+        // Make the path match the os. Debugger seems to return
+        // invalid path chars on linux/darwin
+        if (process.platform !== 'win32') {
+            return path.replace(/\\/g, '/');
+        }
+        return path;
+    }
+
     private isStopEvent(message: DebugProtocol.ProtocolMessage) {
         if (message.type === 'event') {
             const eventMessage = message as DebugProtocol.Event;
@@ -85,37 +103,6 @@ export class DebugLocationTracker extends DebugStackTraceTracker implements Debu
         return false;
     }
 
-    // tslint:disable-next-line:no-any
-    private getStackTrace(message: DebugProtocol.ProtocolMessage): IDebugLocation | undefined {
-        if (message.type === 'response') {
-            const responseMessage = message as DebugProtocol.Response;
-            if (responseMessage.command === 'stackTrace') {
-                const messageBody = responseMessage.body;
-                if (
-                    messageBody.stackFrames.length > 0 &&
-                    this.stackFrameRequestSequenceNumber === responseMessage.request_seq
-                ) {
-                    const lineNumber = messageBody.stackFrames[0].line;
-                    const fileName = this.normalizeFilePath(messageBody.stackFrames[0].source.path);
-                    const column = messageBody.stackFrames[0].column;
-                    return { lineNumber, fileName, column };
-                }
-            }
-        }
-
-        return undefined;
-    }
-
-    private normalizeFilePath(path: string): string {
-        // Make the path match the os. Debugger seems to return
-        // invalid path chars on linux/darwin
-        if (process.platform !== 'win32') {
-            return path.replace(/\\/g, '/');
-        }
-        return path;
-    }
-
-    // tslint:disable-next-line:no-any
     private isContinueEvent(message: DebugProtocol.ProtocolMessage): boolean {
         if (message.type === 'event') {
             const eventMessage = message as DebugProtocol.Event;
@@ -130,5 +117,14 @@ export class DebugLocationTracker extends DebugStackTraceTracker implements Debu
         }
 
         return false;
+    }
+
+    private isResponseForRequestToFetchAllFrames(message: DebugProtocol.Response) {
+        return (
+            message.type === 'response' &&
+            message.command === 'stackTrace' &&
+            message.body.stackFrames[0] &&
+            (message.request_seq === this.stackFrameRequestSequenceNumber || message.request_seq === undefined)
+        );
     }
 }
